+++ chaosaws/__init__.py
@@ -16,7 +16,7 @@ import requests
 from chaosaws.types import AWSResponse
 
 
-__version__ = '0.2.0'
+__version__ = '0.4.0'
 __all__ = ["__version__", "discover", "aws_client", "signed_api_call"]